Chore [#79] 구조체 파일 옮김
