+++ YELLO-iOS/YELLO-iOS/Network/Voting/DTO/Request/VotingAnswerListRequestDTO.swift
@@ -11,29 +11,3 @@ struct VotingAnswerListRequestDTO: Encodable {
     let voteAnswerList: [VoteAnswerList]
     let totalPoint: Int
 }
-
-struct VoteAnswerList: Codable {
-    let friendId: Int
-    let questionId: Int
-    let keywordName: String
-    let colorIndex: Int
-}
-
-// 구조체를 UserDefault에 저장하는 함수
-func saveUserData(_ userData: [VoteAnswerList]) {
-    let encoder = JSONEncoder()
-    if let encodedData = try? encoder.encode(userData) {
-        UserDefaults.standard.set(encodedData, forKey: "UserDataKey")
-    }
-}
-
-// UserDefault에서 구조체를 가져오는 함수
-func loadUserData() -> [VoteAnswerList]? {
-    if let encodedData = UserDefaults.standard.data(forKey: "UserDataKey") {
-        let decoder = JSONDecoder()
-        if let userData = try? decoder.decode([VoteAnswerList].self, from: encodedData) {
-            return userData
-        }
-    }
-    return nil
-}
